extract error utils and add validation check
